API/requirements.txt
alembic==1.13.1
python-dotenv==1.0.0
fastapi==0.105.0
pydantic==2.5.2
pydantic_core==2.14.5
typing_extensions==4.9.0
requests==2.31.0
pandas==2.0.3
passlib==1.7.4
python-jose==3.3.0
python-multipart==0.0.6
Jinja2==3.1.2
starlette==0.27.0
starlette-admin==0.12.2